requirements.txt
fastapi[all]
uvicorn[standard]
dagger-io
pytest
pytest-xdist
ruff
mypy
pre-commit
